scripts/requirements.txt
pandas==1.5.1
PyYAML==6.0
requests==2.28.1
xlrd==2.0.1
